utility/agentConfiguration/pyproject.toml
[project]
name = "agentconfiguration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=2.0.0b3",
    "azure-identity>=1.25.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
